feat: Add Base and BlowUp components
This commit adds the Base and BlowUp components to the project. The Base component is a new file that exports a default Vue component. The BlowUp component is also a new file that exports a default Vue component. These components are added to improve the functionality and enhance the user interface of the application.

Code changes:
- Added Base component
- Added BlowUp component