feat: Add getData and setData methods to cache object
This commit adds the `getData` and `setData` methods to the `cache` object in the `preload.ts` file. The `getData` method invokes the `StoreController:getData` function using `ipcRenderer`, while the `setData` method invokes the `StoreController:setData` function with the provided data. These methods enhance the functionality of the cache object by allowing the retrieval and setting of cache data.

Code changes:
- Added `getData` method to `cache` object
- Added `setData` method to `cache` object